content/base/test/test_websocket_hello.html

<!DOCTYPE HTML>
<html>
<!--
-->
<head>
  <title>Basic websocket test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="testWebSocket()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=472529">Mozilla Bug </a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

var ws;

function testWebSocket () {
  ws = new MozWebSocket("ws://mochi.test:8888/tests/content/base/test/file_websocket_hello");
  ws.onopen = function(e) {
    ws.send("data");
  }
  ws.onclose = function(e) {
  }
  ws.onerror = function(e) {
    ok(false, "onerror called!");
    SimpleTest.finish();
  }
  ws.onmessage = function(e) {
    is(e.data, "Hello world!", "Wrong data");
    ws.close();
    SimpleTest.finish();
  }
}

SimpleTest.waitForExplicitFinish();

</script>
</pre>
<div>


</div>


</body>
</html>
